feat: create LaunchPadDashboardClient component and refactor LaunchPadDashboardPage to utilize it with Suspense for improved client-side rendering
